+++ src/main/java/fast/bank/api/domain/account/service/AccountService.java
@@ -2,9 +2,15 @@ package fast.bank.api.domain.account.service;
 
 import fast.bank.api.domain.account.dto.AccountDetailingData;
 import fast.bank.api.domain.account.dto.AccountRegistrationData;
+import fast.bank.api.domain.account.dto.AccountTransactionRequestData;
+import fast.bank.api.domain.account.dto.AccountTransactionResponseData;
 import fast.bank.api.domain.account.model.Account;
 import fast.bank.api.domain.account.repository.AccountRepository;
 import fast.bank.api.domain.account.service.validation.registration.AccountRegistrationValidators;
+import fast.bank.api.domain.account.service.validation.transaction.AccountTransactionValidators;
+import fast.bank.api.domain.statement.model.Statement;
+import fast.bank.api.domain.statement.model.TransactionType;
+import fast.bank.api.domain.statement.repository.StatementRepository;
 import fast.bank.api.domain.user.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,7 +28,13 @@ public class AccountService {
     private UserRepository userRepository;
 
     @Autowired
+    private StatementRepository statementRepository;
+
+    @Autowired
     private List<AccountRegistrationValidators> registrationValidators;
+
+    @Autowired
+    private List<AccountTransactionValidators> transactionValidators;
 
     @Transactional
     public AccountDetailingData register(AccountRegistrationData data) {
@@ -37,8 +49,21 @@ public class AccountService {
         return new AccountDetailingData(user.getRegistry(), account.getNumber(), account.getAgency(), account.getBalance(), account.getCreditLimit());
     }
 
-    public void transfer() {
+    @Transactional
+    public AccountTransactionResponseData transfer(AccountTransactionRequestData data) {
+        transactionValidators.forEach(v -> v.validate(data));
 
+        var senderAcc = accountRepository.getReferenceById(data.senderAccNumber());
+        var receiverAcc = accountRepository.getReferenceById(data.receiverAccNumber());
+        var transferAmount = data.transferAmount();
+
+        senderAcc.discount(transferAmount);
+        statementRepository.save(new Statement(null, senderAcc, TransactionType.DEBIT, transferAmount, senderAcc.getBalance()));
+        receiverAcc.debit(transferAmount);
+
+        statementRepository.save(new Statement(null, receiverAcc, TransactionType.CREDIT, transferAmount, receiverAcc.getBalance()));
+
+        return new AccountTransactionResponseData(senderAcc.getNumber(), receiverAcc.getNumber(), transferAmount, senderAcc.getBalance());
     }
 
     public void delete() {
